+++ java/client/src/test/java/glide/api/models/configuration/BaseClientConfigurationTest.java
@@ -31,6 +31,11 @@ public class BaseClientConfigurationTest {
             return null;
         }
 
+        @Override
+        public ClientSideCache getClientSideCache() {
+            return null;
+        }
+
         public static class TestClientConfigurationBuilder
                 extends BaseClientConfigurationBuilder<
                         TestClientConfiguration, TestClientConfigurationBuilder> {